Create Tabs for SampleDetails sections (#64)
* Create Tabs for SampleDetails sections

* feat: improve tabs styling

* fix: make tabs work fine with PageContent

* Rename SampleDetails for consistency

* Fix
