+++ frontend/src/styles/antd-adjustments.scss
@@ -5,12 +5,59 @@
  * Distributed under terms of the MIT license.
  */
 
+$darker-border-color: #e0e0e0;
+
+
+.PageContent.--with-tabs {
+  .ant-spin-nested-loading,
+  .ant-spin-container {
+    display: inline-block;
+    overflow: hidden;
+    width: 100%;
+    height: 100%;
+  }
+}
+
+
+/*
+ * Card
+ */
+
 .ant-card .ant-card-head {
   background: #fafafa;
 }
 
 .ant-card-bordered {
-  border: 1px solid #e0e0e0;
+  border: 1px solid $darker-border-color;
+}
+
+/*
+ * Tabs
+ */
+
+.ant-tabs-card > .ant-tabs-nav .ant-tabs-tab,
+.ant-tabs-card > div > .ant-tabs-nav .ant-tabs-tab {
+  border: 1px solid $darker-border-color;
+}
+.ant-tabs-top > .ant-tabs-nav::before,
+.ant-tabs-bottom > .ant-tabs-nav::before,
+.ant-tabs-top > div > .ant-tabs-nav::before,
+.ant-tabs-bottom > div > .ant-tabs-nav::before {
+  border-bottom: 1px solid $darker-border-color;
+}
+
+.ant-tabs {
+  display: inline-flex;
+  width: 100%;
+  height: 100%;
+}
+.ant-tabs-content.ant-tabs-content-top {
+  display: inline-block;
+  max-height: 100%;
+  overflow: auto;
+}
+.ant-tabs-nav-list {
+  padding: 0 8px !important;
 }
